global.dto.pageresponsedto, payment:page 추가, 요청 - post로 변경, dto 분리

+++ backend/src/main/java/com/caffe/global/initdata/BaseInitData.java
@@ -5,7 +5,7 @@ import com.caffe.domain.member.entity.Role;
 import com.caffe.domain.member.repository.MemberRepository;
 import com.caffe.domain.payment.entity.Payment;
 import com.caffe.domain.payment.entity.PaymentOption;
-import com.caffe.domain.payment.entity.PaymentOptionType;
+import com.caffe.global.constant.PaymentOptionType;
 import com.caffe.domain.payment.repository.PaymentOptionRepository;
 import com.caffe.domain.payment.repository.PaymentRepository;
 import com.caffe.domain.product.entity.Product;
@@ -141,13 +141,13 @@ public class BaseInitData {
         Purchase purchase1 = purchaseRepository.findById(1).get();
         PaymentOption paymentOption1 = paymentOptionRepository.findById(4).get();
 
-        Payment payment1 = new Payment((int)purchase1.getTotalPrice(), purchase1, paymentOption1);
+        Payment payment1 = new Payment(purchase1.getTotalPrice(), purchase1, paymentOption1);
         paymentRepository.save(payment1);
 
         Purchase purchase2 = purchaseRepository.findById(2).get();
         PaymentOption paymentOption2 = paymentOptionRepository.findById(7).get();
 
-        Payment payment2 = new Payment((int)purchase2.getTotalPrice(), purchase2, paymentOption2);
+        Payment payment2 = new Payment(purchase2.getTotalPrice(), purchase2, paymentOption2);
         paymentRepository.save(payment2);
 
         System.out.println("결제  데이터 2개  생성 완료");
